Adding Flexible to the sidebar
There was a recent release of Zendesk that allows the bar to be scaled by the user. Adding Flexible to the manifest file will allow the sidebar app to scale with the portion selected by the user. This is extremely valuable for Coveo when viewing documentation that could be larger at times.(forgot 2 commas)